requirements.txt
altgraph==0.17.3
autopep8==2.0.1
beartype==0.11.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
flet==0.3.2
idna==3.4
macholib==1.16.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyNaCl==1.5.0
pyperclip==1.8.2
repath==0.9.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
watchdog==2.2.1
websocket-client==1.4.2